refactor: extract document and test item pipeline helpers
